moved custom-fields settings to extra file and optimized design
